Adição de DTOs e DataAnnotations

+++ Backend/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ProEventos.Application.Dtos;
 using ProEventos.Application.Interfaces;
-using ProEventos.Domain;
-using ProEventos.Persistence.Context;
 
 namespace ProEventos.API.Controllers
 {
@@ -26,7 +23,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await this.eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -43,7 +40,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await this.eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento não encontrado");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -60,7 +57,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await this.eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -72,12 +69,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await this.eventoService.AddEvento(model);
-                if (evento == null) return BadRequest("Erro ao adicionar evento");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -89,12 +86,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await this.eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao atualizar evento");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -110,9 +107,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await this.eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await this.eventoService.DeleteEvento(id) ? 
                     Ok("Evento excluído") : 
-                    BadRequest("Evento não excluído");
+                    throw new Exception("Ocorreu algum problema ao tentar deletar o Evento.");
             }
             catch (Exception ex)
             {
